create group 7 and split e2e json rpc tests
