+++ ci/e2e-group4/jsonrpc1_test.go
@@ -1,0 +1,1 @@
+../../test/e2e/jsonrpc1_test.go
