--- ci/e2e-group4/jsonrpc_test.go
@@ -1,1 +1,0 @@
-../../test/e2e/jsonrpc_test.go
