+++ ci/e2e-group7/jsonrpc2_test.go
@@ -1,0 +1,1 @@
+../../test/e2e/jsonrpc2_test.go
