+++ test/e2e/shared.go
@@ -2,10 +2,27 @@
 package e2e
 
 import (
+	"context"
+	"fmt"
+	"math/big"
+
 	"github.com/0xPolygonHermez/zkevm-node/log"
 	"github.com/0xPolygonHermez/zkevm-node/state"
 	"github.com/0xPolygonHermez/zkevm-node/test/operations"
-	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum"
+	"github.com/ethereum/go-ethereum/accounts/abi/bind"
+	"github.com/ethereum/go-ethereum/common"
+	ethTypes "github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/ethclient"
+)
+
+const (
+	invalidParamsErrorCode = -32602
+	toAddressHex           = "0x4d5Cf5032B2a844602278b01199ED191A86c93ff"
+)
+
+var (
+	toAddress = common.HexToAddress(toAddressHex)
 )
 
 var networks = []struct {
@@ -31,7 +48,65 @@ var networks = []struct {
 	},
 }
 
-func logTx(tx *types.Transaction) {
+func Setup() {
+	var err error
+	ctx := context.Background()
+	err = operations.Teardown()
+	if err != nil {
+		panic(err)
+	}
+
+	opsCfg := operations.GetDefaultOperationsConfig()
+	opsMan, err := operations.NewManager(ctx, opsCfg)
+	if err != nil {
+		panic(err)
+	}
+	err = opsMan.Setup()
+	if err != nil {
+		panic(err)
+	}
+}
+
+func Teardown() {
+	err := operations.Teardown()
+	if err != nil {
+		panic(err)
+	}
+}
+
+func createTX(client *ethclient.Client, auth *bind.TransactOpts, to common.Address, amount *big.Int) (*ethTypes.Transaction, error) {
+	nonce, err := client.NonceAt(context.Background(), auth.From, nil)
+	if err != nil {
+		return nil, err
+	}
+	gasLimit, err := client.EstimateGas(context.Background(), ethereum.CallMsg{From: auth.From, To: &to, Value: amount})
+	if err != nil {
+		return nil, err
+	}
+
+	gasPrice, err := client.SuggestGasPrice(context.Background())
+	if err != nil {
+		return nil, err
+	}
+
+	log.Infof("\nTX details:\n\tNonce:    %d\n\tGasLimit: %d\n\tGasPrice: %d", nonce, gasLimit, gasPrice)
+	if gasLimit != uint64(21000) {
+		return nil, fmt.Errorf("gasLimit %d != 21000", gasLimit)
+	}
+	tx := ethTypes.NewTransaction(nonce, to, amount, gasLimit, gasPrice, nil)
+	signedTx, err := auth.Signer(auth.From, tx)
+	if err != nil {
+		return nil, err
+	}
+	log.Infof("Sending Tx %v Nonce %v", signedTx.Hash(), signedTx.Nonce())
+	err = client.SendTransaction(context.Background(), signedTx)
+	if err != nil {
+		return nil, err
+	}
+	return signedTx, nil
+}
+
+func logTx(tx *ethTypes.Transaction) {
 	sender, _ := state.GetSender(*tx)
 	log.Debugf("********************")
 	log.Debugf("Hash: %v", tx.Hash())
